Split into DraggableContents and DroppableContents components for DragAndDropContent
