+++ src/components/admin/product-category/DragAndDropContent.tsx
@@ -1,50 +1,18 @@
-import styled from '@emotion/styled';
 import useAdminProducts from '@hooks/admin/useAdminProducts';
 import { categoriesAtom } from '@recoils/atoms';
-import DeleteButtonGraySvg from '@resources/svg/DeleteButtonGraySvg';
-import DragIconSvg from '@resources/svg/DragIconSvg';
 import { useEffect } from 'react';
-import { DragDropContext, Draggable, Droppable, DropResult } from 'react-beautiful-dnd';
+import { DragDropContext, Droppable, DropResult } from 'react-beautiful-dnd';
 import { useParams } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import useConfirm from '@hooks/useConfirm';
-import { rowFlex } from '@styles/flexStyles';
-
-const CategoriesItemContainer = styled.div`
-  position: relative;
-  width: 600px;
-  ${rowFlex({ justify: 'center', align: 'center' })}
-`;
-
-const CategoriesContentsContainer = styled.div`
-  width: 500px;
-  height: 60px;
-  background-color: white;
-  border-radius: 12px;
-  box-shadow: 0 4px 17px 0 rgba(0, 0, 0, 0.1);
-  margin: 10px 0;
-  ${rowFlex({ justify: 'space-between', align: 'center' })}
-`;
-
-const CategoriesName = styled.label`
-  font-size: 20px;
-  width: 50%;
-  padding-left: 20px;
-`;
-
-const DeleteIcon = styled(DeleteButtonGraySvg)`
-  position: absolute;
-  left: 13px;
-  &:hover {
-    transform: scale(1.2);
-  }
-`;
+import DroppableContents from './DroppableContents';
+import { ProductCategory } from '@@types/index';
 
 function DragAndDropContent() {
   const { workspaceId } = useParams<{ workspaceId: string }>();
-  const { fetchCategories, deleteCategory } = useAdminProducts(workspaceId);
+  const { fetchCategories } = useAdminProducts(workspaceId);
   const [rawCategories, setRawCategories] = useRecoilState(categoriesAtom);
-  const categories = rawCategories.map((category) => ({
+  const categories: ProductCategory[] = rawCategories.map((category) => ({
     ...category,
     id: String(category.id),
   }));
@@ -55,7 +23,7 @@ function DragAndDropContent() {
     fetchCategories();
   }, []);
 
-  const reorder = (list: any[], startIndex: number, endIndex: number) => {
+  const reorder = (list: ProductCategory[], startIndex: number, endIndex: number) => {
     const result = Array.from(list);
     const [removed] = result.splice(startIndex, 1);
 
@@ -75,34 +43,7 @@ function DragAndDropContent() {
 
   return (
     <DragDropContext onDragEnd={onDragEnd}>
-      <Droppable droppableId="droppable">
-        {(provided) => (
-          <div ref={provided.innerRef} {...provided.droppableProps}>
-            {categories.map((item, index) => (
-              <Draggable key={item.id} draggableId={item.id} index={index}>
-                {(pro) => (
-                  <CategoriesItemContainer ref={pro.innerRef} {...pro.draggableProps} {...pro.dragHandleProps}>
-                    <DeleteIcon
-                      onClick={() => {
-                        deleteCategory(Number(item.id)).catch((e) => {
-                          if (e.response.status === 405) {
-                            confirm();
-                          }
-                        });
-                      }}
-                    />
-                    <CategoriesContentsContainer>
-                      <CategoriesName>{item.name}</CategoriesName>
-                      <DragIconSvg style={{ paddingRight: '20px' }} />
-                    </CategoriesContentsContainer>
-                  </CategoriesItemContainer>
-                )}
-              </Draggable>
-            ))}
-            {provided.placeholder}
-          </div>
-        )}
-      </Droppable>
+      <Droppable droppableId="droppable">{(provided) => <DroppableContents provided={provided} categories={categories} confirm={confirm} />}</Droppable>
       <ConfirmModal />
     </DragDropContext>
   );
